third_party/WebKit/LayoutTests/virtual/color_space/fast/canvas/color-space/canvas-drawImage-p3.html

<!DOCTYPE HTML>
<script src="../../../../../resources/testharness.js"></script>
<script src="../../../../../resources/testharnessreport.js"></script>
<script>

var p3Red = [0.822, 0.033, 0.017, 1];
var p3Green = [0.177, 0.966, 0.072, 1];
var p3Blue = [0, 0, 0.910, 1];
var p3Black = [0, 0, 0, 1];

function testPixels(ctx, tests)
{
    var actual, expected, tolerance = 0.01;
    for (var i = 0; i < tests.length; i++) {
      actual = ctx.getImageData(tests[i][0], tests[i][1], 1, 1).dataUnion;
      expected = tests[i][2];
      assert_true(actual.length === expected.length);
      for (var i = 0; i < actual.length; i++)
        assert_approx_equals(actual[i], expected[i], tolerance);
    }
}

function drawSRGBImageOnP3Canvas(source)
{
  var canvas = document.createElement('canvas');
  canvas.width = 20;
  canvas.height = 20;
  var ctx = canvas.getContext('2d',
      {colorSpace: 'p3', pixelFormat:'float16'});
  ctx.drawImage(source, 0, 0);
  var tests = [[5, 5, p3Red], 
               [15, 5, p3Green], 
               [5, 15, p3Blue], 
               [15, 15, p3Black]];
  testPixels(ctx, tests);
}

promise_test(function() {
  return new Promise((resolve, reject) => {
    var image = new Image();
    image.onload = function() {
      resolve(image);
    }
    image.src = 'resources/pattern-srgb.png'
  }).then(drawSRGBImageOnP3Canvas);
}, 'Draw SRGB image on a P3 canvas and read back the P3 pixels.');

// TODO(zakerinasab): crbug.com/713867
// Add at least two more tests to load pattern-p3d65.png and pattern-rec2020.png
// and draw them on a P3 canvas.

</script>
